調整 speaker 架構
