+++ 2019-dev/pages/speaker/images/icon/icon-github.png
